server/requirements.txt
openenv-core
pydantic>=2.0,<3.0
fastapi
uvicorn[standard]
pyyaml>=6.0
